Custom header file and separate module for string functions
